refactor: remove layout component and integrate server-side session handling in BracketsPage
- Deleted BracketsLayout component and its associated session provider.
- Refactored BracketsPage to use server-side session retrieval with getServerAuthSession.
- Simplified the component structure by directly rendering BracketsClient based on user role.
- Removed unused types and data structures related to beginners and advanced brackets.
